pyproject.toml
[tool.poetry]
name = "research-project"
version = "0.1.0"
description = "Economic analysis ingestion scaffolding"
authors = []
readme = "README.md"
packages = [{ include = "ingestion" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pydantic = "^2.11.0"
pandas = "^2.2.0"
pyarrow = "^19.0.0"
httpx = "^0.28.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
ruff = "^0.11.0"

[tool.poetry.scripts]
ingestion-cli = "ingestion.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
  "integration: opt-in live endpoint tests for raw ingestion sanity checks",
]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
